Tried to add a shadow map

+++ Settings.cpp
@@ -3,6 +3,6 @@ float Settings::DOWNSCALING = 1.0f;
 bool Settings::ADVANCED_WATER = true;
 bool Settings::SKYBOX = true;
 bool Settings::AMBIENT_OCCLUSION = true;
-int Settings::FOV = 90;
+float Settings::FOV = 90.0f;
 int Settings::RENDER_DISTANCE = 10;
 bool Settings::BLOOM = false;
