+++ Settings.h
@@ -4,7 +4,7 @@ public:
 	static bool ADVANCED_WATER;
 	static bool SKYBOX;
 	static bool AMBIENT_OCCLUSION;
-	static int FOV;
+	static float FOV;
 	static int RENDER_DISTANCE;
 	static float DOWNSCALING;
 	static bool BLOOM;
